Thehyve

+++ grails-app/domain/org/transmartproject/db/dataquery/highdim/snp_lz/DeSnpInfo.groovy
@@ -4,7 +4,7 @@ class DeSnpInfo {
 
     String name
     String chromosome
-    Integer pos
+    Long pos
 
     static mapping = {
         table      schema: 'deapp',       name:      'de_snp_info'
